% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Load the comma-separated data files in a directory.}
\usage{
compile(data_dir, regex_key = "*")
}
\arguments{
\item{data_dir}{The directory of the data files, passed to `file.path`.}

\item{regex_key}{Pattern passed to `list.files`. Matches are loaded.}
}
\value{
dplyr::data_frame
}
\description{
Load the comma-separated data files in a directory.
}

