% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{format_mod_preds}
\alias{format_mod_preds}
\title{Tidy up the model predictions for plotting.}
\usage{
format_mod_preds(mod, response_type = c("pic", "prompt"))
}
\arguments{
\item{mod}{Model (lmer, glmer) to pass to predict function.}

\item{response_type}{pic or prompt.}
}
\value{
a tidy data frame containing all x predictors and y predictions.
}
\description{
X predictions are all unique combinations of
feat_type (visual, nonvisual), mask_type (nomask, mask),
and response_type (pic, prompt). Character combinations
are sent to the recode function in this package.
}

