% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode all of the factors in the experiment.}
\usage{
recode(frame)
}
\arguments{
\item{frame}{A data.frame with columns: feat_type, mask_type, response_type}
}
\description{
Recode all of the factors in the experiment.
}

